Move Metadata & Protocol back to Client
