--- src/main/scala/scalaz/distributed/Protocol.scala
@@ -1,5 +1,0 @@
-package scalaz.distributed
-
-trait Protocol[Type[_], Path[_, _]] {
-  def access[A: Type](id: MetadataID): F[Metadata[Type, Path]]
-}
